+++ libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: ohakola <ohakola@student.hive.fi>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/15 14:42:13 by ohakola           #+#    #+#             */
-/*   Updated: 2020/08/26 22:08:43 by ohakola          ###   ########.fr       */
+/*   Updated: 2020/08/26 23:14:08 by ohakola          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -58,10 +58,10 @@ typedef struct	s_pixel_bounds
 }				t_pixel_bounds;
 
 typedef struct	s_float_bits {
-	unsigned long long int	fraction:63;
-	unsigned long long int	intbit:1;
-	unsigned int			exp:15;
-	unsigned int			sign:1;
+	unsigned long long int		fraction:63;
+	unsigned long long int		intbit:1;
+	unsigned long long int		exp:15;
+	unsigned long long int		sign:1;
 }				t_float_bits;
 
 typedef union	u_float_dissector {
